% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reeval_efunctions.R
\name{reeval_efunctions}
\alias{reeval_efunctions}
\title{Get new spline basis for eigenfunctions}
\usage{
reeval_efunctions(knots, argvals_bin, argvals, efunctions, npc)
}
\arguments{
\item{knots}{Number of knots for spline basis}

\item{argvals_bin}{Index on for binned data.}

\item{argvals}{Full index on which to generate new basis.}

\item{efunctions}{Eigenfunctions from an fpca object.}

\item{npc}{Number of principal components.}
}
\description{
Internal function for generating spline basis from fpca.face results to evaluate eigenfunctions
on expanded grid.
}
