% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gfpca.R
\name{sim_gfpca}
\alias{sim_gfpca}
\title{Simulate GFPCA}
\usage{
sim_gfpca(N = 500, J = 100, case = 1, family = "binomial")
}
\arguments{
\item{N}{Number of subjects.}

\item{J}{Number of timepoints per subject.}

\item{case}{Takes on values of 1 or 2. If case = 1 true eigenfunctions
are based on alternating sine and cosines. If case = 2 true eigenfunctions are
based on sqrt functions.}

\item{family}{not currently implemented. Currently returns binary data. Will be adapted for
Poisson and Gaussian families as well.}
}
\description{
This function generates simulated generalized functional principal components data.
Simulation setup is adapted from code in the help file for the function fpca.face
from the refund package. 4 true principal components are generated.
}
