% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_gfpca.R
\name{fast_gfpca}
\alias{fast_gfpca}
\title{Fast generalized functional principal components analysis}
\usage{
fast_gfpca(
  Y,
  argvals = NULL,
  overlap = FALSE,
  binwidth = 10,
  pve = NULL,
  npc = NULL,
  family = "binomial",
  periodicity = FALSE,
  ...
)
}
\arguments{
\item{Y}{Dataframe. Should have variables id, value, index.}

\item{argvals}{numeric; grid over which functions are observed.  If null defaults to unique values of index.}

\item{overlap}{Logical; indicates whether or not to construct overlapping bins. Defaults to FALSE}

\item{binwidth}{controls the width of the bins for step 1. Must have an even integer value (defaults to 10). Can be no more than J/10.}

\item{pve}{proportion of variance explained: used to choose the number of
principal components unless \code{npc} is specified.}

\item{npc}{how many smooth PCs to try to extract, if \code{NULL} (the
default) then npc is chosen based on \code{pve}}

\item{family}{exponential family to be passed to \code{glmer} and \code{bam}.}

\item{periodicity}{Option for a periodic spline basis. Defaults to FALSE.}

\item{...}{Additional arguments passed to or from other functions}
}
\value{
An object of class \code{fpca} containing:
\item{efunctions}{\eqn{D \times npc} matrix of estimated FPC basis functions.}
\item{evalues}{Estimated variance of the FPC scores.}
\item{npc}{number of FPCs.}
\item{scores}{\eqn{I \times npc} matrix of estimated FPC scores.}
\item{mu}{Estimated population-level mean.}
\item{Yhat}{FPC approximation of subject-specific means, before applying the
response function.}
\item{Y}{The observed data.}
\item{family}{for compatibility with \code{refund.shiny} package.}
}
\description{
This is the main function for the fastGFPCA package. The function requires input data \code{Y} to be a dataframe in long format with variables
\code{id}, \code{index}, and \code{value} to indicate subject IDs,
observation times on the domain, and observations, respectively.
The \code{index} must contain the same, equally spaced grid points for each subject.
The number of functional principal components (FPCs) can either be specified
directly (argument \code{npc}) or chosen based on the explained share of
variance (\code{pve}). Families and link functions can be specified using the same
syntax as the \code{family} argument from the \code{lme4::glmer()} function.
}
\examples{
# simulate data
set.seed(1001)

# binomial data, with bins that do not overlap
df_gfpca <- sim_gfpca(N = 200, J = 200, case = 2)$df_gfpca
gfpca_mod <- fast_gfpca(df_gfpca, overlap = FALSE, binwidth = 10, family = binomial)

# Poisson data, overlapping bins
df_gfpca <- sim_gfpca(N = 100, J = 100, case = 1, family = "poisson")$df_gfpca
gfpca_mod <- fast_gfpca(df_gfpca, overlap = TRUE, binwidth =6, family = "poisson")

}
\references{
Xiao, L., Ruppert, D., Zipunnikov, V., and Crainiceanu, C. (2016).
Fast covariance estimation for high-dimensional functional data.
\emph{Statistics and Computing}, 26, 409-421.
DOI: 10.1007/s11222-014-9485-x.
}
\author{
Andrew Leroux \email{andrew.leroux@cuanschutz.edu},
Julia Wrobel \email{julia.wrobel@cuanschutz.edu}
}
