% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gfpca.R
\name{sim_gfpca}
\alias{sim_gfpca}
\title{Simulate GFPCA}
\usage{
sim_gfpca(
  N = 500,
  J = 100,
  case = 1,
  mu = FALSE,
  family = "binomial",
  sigma = 2
)
}
\arguments{
\item{N}{Number of subjects.}

\item{J}{Number of timepoints per subject.}

\item{case}{Takes on values of 1 or 2. If case = 1 true eigenfunctions
are based on alternating sine and cosines. If case = 2 true eigenfunctions are
based on sqrt functions.}

\item{mu}{Logical to indicate if a mean function should be included in the simulated data. Defaults to false (mu = 0).
If TRUE, a mean function is constructed using B-splines.}

\item{family}{Defines exponential family for generating data. Options are binomial (default),
poisson, or gaussian.}

\item{sigma}{Defaults to 2. Error variance for Y when family = "gaussian".}
}
\description{
This function generates simulated generalized functional principal components data.
Simulation setup is adapted from code in the help file for the function fpca.face
from the refund package. 4 true principal components are generated.
}
\examples{
# generate binary data
df <- sim_gfpca(N = 200, J = 100, case = 1, family = "binomial")$df_gfpca

# generate Poisson data
df <- sim_gfpca(N = 200, J = 100, case = 1, family = "poisson")$df_gfpca

# generate Gaussian data
df <- sim_gfpca(N = 200, J = 100, case = 1, family = "gaussian")$df_gfpca

}
